pyproject.toml
[project]
name = "webview-boilerplate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jinja2-simple-tags>=0.6.1",
    "jinja2>=3.1.4",
    "psutil>=6.1.0",
    "pyperclip>=1.9.0",
    "pywebview>=5.3.2",
]

[dependency-groups]
dev = [
    "deep-translator>=1.11.4",
    "isort>=5.13.2",
    "pyinstaller>=6.11.1",
    "watchfiles>=0.24.0",
]


[tool.ruff.format]
quote-style = 'single'


[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PT']
